Make assertRaises / assertWarns context managers work
Fixes #4. Note that since the semantics of what e.g. `assertRaises` captures is different to what `pytest.raises` captures, they can't easily be converted. It's recommended users manually fix up any cases after the fact - added a note to the README to indicate this.

Also fixed the link to pytest in the README, it was pointing to PEP8.